requirements.txt
# Core Discord Bot
discord.py==2.4.0
python-dotenv==1.0.0

# Audio & Music Streaming
wavelink==3.3.0
yt-dlp==2024.1.1
spotipy==2.22.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# HTTP & APIs
aiohttp==3.9.1
requests==2.31.0

# Data Validation
pydantic==2.5.0

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.39.0

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.23.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
